Methods

Methods

This section describes systems methods implemented in HybridSystems.jl.

Methods

Hybrid automata

Hybrid automata

HybridSystems.states — Function.

states(A::AbstractAutomaton)

Returns an iterator over the states of the automaton A. It has the alias modes.

source

HybridSystems.nstates — Function.

nstates(A::AbstractAutomaton)

Returns the number of states of the automaton A. It has the alias nmodes.

source

HybridSystems.transitiontype — Function.

transitiontype(A::AbstractAutomaton)

Returns type of the transitions of the automaton A.

source

HybridSystems.transitions — Function.

transitions(A::AbstractAutomaton)

Returns an iterator over the transitions of the automaton A.

source

HybridSystems.ntransitions — Function.

ntransitions(A::AbstractAutomaton)

Returns the number of transitions of the automaton A.

source

HybridSystems.add_transition! — Function.

add_transition!(A::AbstractAutomaton, q, r, σ)

Adds a transition between states q and r with symbol σ to the automaton A.

source

HybridSystems.has_transition — Function.

has_transition(A::AbstractAutomaton, t)

Returns true if the automaton A has the transition t.

source

HybridSystems.rem_transition! — Function.

rem_transition!(A::AbstractAutomaton, q, r, σ)

Remove the transition between states q and r with symbol σ to the automaton A.

source

HybridSystems.rem_state! — Function.

rem_state!(A::AbstractAutomaton, q)

Remove the state q to the automaton A.

source

HybridSystems.source — Function.

source(A::AbstractAutomaton, t)

Returns the source of the transition t.

source

HybridSystems.event — Function.

event(A::AbstractAutomaton, t)

Returns the event/symbol of the transition t in the automaton A. It has the alias symbol.

source

HybridSystems.target — Function.

target(A::AbstractAutomaton, t)

Returns the target of the transition t.

source

HybridSystems.in_transitions — Function.

in_transitions(A::AbstractAutomaton, s)

Returns an iterator over the transitions with target s.

source

HybridSystems.out_transitions — Function.

out_transitions(A::AbstractAutomaton, s)

Returns an iterator over the transitions with source s.

source