Methods

Methods

This section describes systems methods implemented in HybridSystems.jl.

Methods

Switched Systems

Continuous sub-systems

Hybrid automata

Modes

Transitions

Switched Systems

The following method makes it easy to create specific kind of hybrid systems called switched systems

HybridSystems.discreteswitchedsystem — Function.

discreteswitchedsystem(A::AbstractVector{<:AbstractMatrix})

Creates the discrete switched linear system defined by

\[x_{k+1} = A_{\sigma_k} x_k, \sigma_k = 1, \ldots, m.\]

where m is the length of A.

discreteswitchedsystem(A::AbstractVector{<:AbstractMatrix}, S::AbstractVector)

Creates the state dependent discrete switched linear system defined by

\[x_{k+1} = A_{\sigma_k} x_k, \sigma_k = 1, \ldots, m, x_k \in S[\sigma_k].\]

where m is the length of A and S.

discreteswitchedsystem(A::AbstractVector{<:AbstractMatrix}, G::AbstractAutomaton)

Creates the constrained discrete switched linear system defined by

\[x_{k+1} = A_{\sigma_k} x_k,\]

where $\sigma_1, \ldots, \sigma_k$ is a valid sequence of events of the automaton G.

discreteswitchedsystem(A::AbstractVector{<:AbstractMatrix}, G::AbstractAutomaton, S::AbstractVector)

Creates the state-dependent constrained discrete switched linear system defined by

\[x_{k+1} = A_{\sigma_k} x_k, x_k \in S[q_k]\]

where $q_0, \sigma_1, q_1, \ldots, q_{k-1}, \sigma_k, q_k$ is a valid sequence of events of the automaton G with intermediate states $q_0, \ldots, q_k$.

source

Continuous sub-systems

MathematicalSystems.statedim — Function.

statedim(hs::HybridSystem, u::Int)

Returns the dimension of the state space of the system at mode u.

source

MathematicalSystems.stateset — Function.

stateset(s::AbstractSystem, u::Int)

Returns the set of allowed states of the system at mode u.

source

stateset(s::AbstractSystem, t)

Returns the guard for the transition t.

source

MathematicalSystems.inputdim — Function.

inputdim(s::AbstractSystem, u::Int)

Returns the dimension of the input space of the system at mode u.

source

MathematicalSystems.inputset — Function.

inputset(s::AbstractSystem, u::Int)

Returns the set of allowed inputs of the system at mode u.

source

inputset(s::AbstractSystem, t)

Returns the st of allowed inputs for the transition t.

source

Hybrid automata

Modes

HybridSystems.states — Function.

states(A::AbstractAutomaton)

Returns an iterator over the states of the automaton A. It has the alias modes.

source

HybridSystems.nstates — Function.

nstates(A::AbstractAutomaton)

Returns the number of states of the automaton A. It has the alias nmodes.

source

HybridSystems.rem_state! — Function.

rem_state!(A::AbstractAutomaton, q)

Remove the state q to the automaton A.

source

Transitions

HybridSystems.transitiontype — Function.

transitiontype(A::AbstractAutomaton)

Returns type of the transitions of the automaton A.

source

HybridSystems.transitions — Function.

transitions(A::AbstractAutomaton)

Returns an iterator over the transitions of the automaton A.

source

HybridSystems.ntransitions — Function.

ntransitions(A::AbstractAutomaton)

Returns the number of transitions of the automaton A.

source

HybridSystems.add_transition! — Function.

add_transition!(A::AbstractAutomaton, q, r, σ)

Adds a transition between states q and r with symbol σ to the automaton A.

source

HybridSystems.has_transition — Function.

has_transition(A::AbstractAutomaton, t)

Returns true if the automaton A has the transition t.

source

HybridSystems.rem_transition! — Function.

rem_transition!(A::AbstractAutomaton, q, r, σ)

Remove the transition between states q and r with symbol σ to the automaton A.

source

HybridSystems.source — Function.

source(A::AbstractAutomaton, t)

Returns the source of the transition t.

source

HybridSystems.event — Function.

event(A::AbstractAutomaton, t)

Returns the event/symbol of the transition t in the automaton A. It has the alias symbol.

source

HybridSystems.target — Function.

target(A::AbstractAutomaton, t)

Returns the target of the transition t.

source

HybridSystems.in_transitions — Function.

in_transitions(A::AbstractAutomaton, s)

Returns an iterator over the transitions with target s.

source

HybridSystems.out_transitions — Function.

out_transitions(A::AbstractAutomaton, s)

Returns an iterator over the transitions with source s.

source